Move _unstack_input_sequence from recurrent to recurrent_internals
